requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-bootstrap==0.2.4
django-bootstrap4==2.1.1
django-crispy-forms==1.9.1
django-fontawesome==1.0
google==2.0.3
google-auth==1.18.0
idna==2.9
kubernetes==11.0.0
oauthlib==3.1.0
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylogrus==0.4.0
python-bootstrap==0.0.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
shortuuid==1.0.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
websocket-client==0.57.0
<<<<<<< HEAD
=======

django-filter~=2.3.0
>>>>>>> aa3d5307... update models.py
